pyproject.toml
[project]
name = "proxysub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.127.0",
    "pyyaml>=6.0.3",
    "uvicorn>=0.30.0",
]
